Finally a Plan update!
From now on the registration process, users can choose to use a customAlertDialog who open some question toward the user and from there the code generate for the user three different Plans in which the user can pick from.

The new customAlertDialog also needed to be added with a brand new feature who let the user edit his Plan from the UserInfoScreen !

A lot of mini-changes and cleaning added.